requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
blas=1.0=mkl
blosc2=2.0.0=pypi_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
bzip2=1.0.8=h7b6447c_0
ca-certificates=2023.01.10=h06a4308_0
certifi=2022.12.7=py310h06a4308_0
contourpy=1.0.5=py310hdb19cb5_0
cycler=0.11.0=pyhd3eb1b0_0
cython=0.29.34=pypi_0
dbus=1.13.18=hb2f20db_0
expat=2.4.9=h6a678d5_0
fontconfig=2.14.1=h52c9d5c_1
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
giflib=5.2.1=h5eee18b_3
glib=2.69.1=he621ea3_2
gst-plugins-base=1.14.1=h6a678d5_1
gstreamer=1.14.1=h5eee18b_1
icu=58.2=he6710b0_3
intel-openmp=2021.4.0=h06a4308_3561
jpeg=9e=h5eee18b_1
kiwisolver=1.4.4=py310h6a678d5_0
krb5=1.19.4=h568e23c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libclang=10.0.1=default_hb85057a_2
libdeflate=1.17=h5eee18b_0
libedit=3.1.20221030=h5eee18b_0
libevent=2.1.12=h8f2d780_0
libffi=3.4.2=h6a678d5_6
libgcc-ng=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libllvm10=10.0.1=hbcb73fb_5
libpng=1.6.39=h5eee18b_0
libpq=12.9=h16c4e8d_3
libstdcxx-ng=11.2.0=h1234567_1
libtiff=4.5.0=h6a678d5_2
libuuid=1.41.5=h5eee18b_0
libwebp=1.2.4=h11a3e52_1
libwebp-base=1.2.4=h5eee18b_1
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lz4-c=1.9.4=h6a678d5_0
matplotlib=3.7.1=py310h06a4308_1
matplotlib-base=3.7.1=py310h1128e8f_1
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
msgpack=1.0.5=pypi_0
munkres=1.1.4=py_0
ncurses=6.4=h6a678d5_0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numexpr=2.8.4=pypi_0
numpy=1.23.5=py310hd5efca6_0
numpy-base=1.23.5=py310h8e6c178_0
openssl=1.1.1t=h7f8727e_0
packaging=23.0=py310h06a4308_0
pcre=8.45=h295c915_0
pillow=9.4.0=py310h6a678d5_0
pip=23.0.1=py310h06a4308_0
ply=3.11=py310h06a4308_0
py-cpuinfo=9.0.0=pypi_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
python=3.10.10=h7a1cb2a_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python-unsio=1.0.1=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
readline=8.2=h5eee18b_0
setuptools=65.6.3=py310h06a4308_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
sqlite=3.41.1=h5eee18b_0
tables=3.8.0=pypi_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
tornado=6.2=py310h5eee18b_0
tzdata=2022g=h04d1e81_0
wheel=0.38.4=py310h06a4308_0
xz=5.2.10=h5eee18b_1
zlib=1.2.13=h5eee18b_0
zstd=1.5.4=hc292b87_0

numpy~=1.23.5
tables~=3.8.0